+++ src/main/java/web/mvc/domain/QaBoard.java
@@ -17,7 +17,7 @@ public class QaBoard {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "board_no_seq")
-    private Integer boardNoSeq;
+    private Long boardNoSeq;
 
     @Column(name = "subject", nullable = false, length = 45)
     private String subject;
