+++ src/main/java/web/mvc/dto/AuctionDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @Builder
 @AllArgsConstructor
 public class AuctionDTO {
-    private int auctionSeq;
+    private Long auctionSeq;
     private int count;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
